feat(broker): 实现消费队列功能
- 新增消费队列相关模型和处理逻辑
- 实现消费队列的内存映射文件加载和写入
- 添加消费组和偏移量管理功能
- 优化主题配置文件读写

+++ mq-broker/src/main/java/org/example/broker/model/TopicLatestCommitLogModel.java
@@ -2,7 +2,7 @@ package org.example.broker.model;
 
 import lombok.Data;
 
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * @author qushutao
@@ -15,9 +15,9 @@ public class TopicLatestCommitLogModel {
     private String fileName;
 
     // 文件最大偏移量
-    private long offsetLimit;
+    private int offsetLimit;
 
     // 偏移量最新写入文件的位置
-    private AtomicLong offset;
+    private AtomicInteger offset;
 
 }
